requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
scikit-learn==1.3.0
opencv-python==4.8.0.74
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
pyfingerprint==0.4